requirements.txt
# Desktop shell dependencies for the native app window.
pywebview>=5.1
PyQt5>=5.15
PyQtWebEngine>=5.15
qtpy>=2.4
